move UpgradeClusterComponentData into separate file
