+++ src/app/api/model/UpgradeClusterDialogData.ts
@@ -1,0 +1,10 @@
+export class UpgradeClusterComponentData {
+  clusterName: string;
+  upgradesList: string[];
+
+  constructor(clusterName: string, upgradesList: string[]) {
+    this.clusterName = clusterName;
+    this.upgradesList = upgradesList;
+  }
+
+}
